app/requirements.txt
prometheus_client==0.20.0
requests==2.31.0
schedule==1.2.2
lxml==5.3.0
pyyaml==6.0.1